// src/app/models/ariziModels/ariziGrup.ts
export class AriziGrup {
  Id: number;
  UetdsSeferReferansNo: number;
  UetdsGrupRefNo: number;
  GrupAdi: string;
  GrupAciklama: string;
  GrupUcret: string;
  BaslangicUlke: string;
  BaslangicIl: number;
  BaslangicIlce: number;
  BaslangicYer: string;
  BitisUlke: string;
  BitisIl: number;
  BitisIlce: number;
  BitisYer: string;
  AktifIptal: string;
  IptalAciklama: string;
}