// src/app/models/ariziModels/ariziGrupDto.ts
export interface AriziGrupDto {
  Id: number;
  UetdsSeferReferansNo: number;
  UetdsGrupRefNo: number;
  GrupAdi: string;
  GrupAciklama: string;
  GrupUcret: string;
  BaslangicUlke: string;
  BaslangicIl: number;
  BaslangicIlAdi: string;
  BaslangicIlce: number;
  BaslangicIlceAdi: string;
  BaslangicYer: string;
  BitisUlke: string;
  BitisIl: number;
  BitisIlAdi: string;
  BitisIlce: number;
  BitisIlceAdi: string;
  BitisYer: string;
  AktifIptal: string;
  IptalAciklama: string;
}